Refactored submit tests
